pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vcon-telephony-adapters"
version = "0.2.0"
description = "Telephony adapters for converting recordings to vCon format - supports Twilio and more"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "vCon Community"}
]
keywords = ["vcon", "twilio", "adapter", "conserver", "telephony", "recordings", "freeswitch", "asterisk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Telephony",
    "Topic :: Multimedia :: Sound/Audio",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "twilio>=8.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "vcon-telephony-adapters[dev]",
]

[project.scripts]
vcon-adapter = "main:main"
# Keep backwards compatibility
vcon-twilio-adapter = "main:main"

[project.urls]
Homepage = "https://github.com/vcon-dev/vcon-telephony-adapters"
Repository = "https://github.com/vcon-dev/vcon-telephony-adapters"
Issues = "https://github.com/vcon-dev/vcon-telephony-adapters/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "adapters*", "twilio_adapter*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = [
    "E501",   # Line too long (handled by black)
    "N803",   # Argument name should be lowercase (Twilio uses PascalCase params)
    "B008",   # Do not perform function call in argument defaults (FastAPI pattern)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["N803", "B008"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=core --cov=adapters --cov=twilio_adapter --cov-report=term-missing"
asyncio_mode = "auto"
